Updated container memory report. From RSS+cache to RSS memory
