+++ code/agent/activate.py
@@ -27,6 +27,7 @@ class Activate(NuvlaBoxCommon.NuvlaBoxCommon):
         # self.activation_flag = "{}/.activated".format(self.data_volume)
         super().__init__(shared_data_volume=data_volume)
 
+        self.activate_logger: logging.Logger = logging.getLogger(__name__)
         # self.api = nb.ss_api() if not api else api
         self.user_info = {}
 
@@ -43,15 +44,15 @@ class Activate(NuvlaBoxCommon.NuvlaBoxCommon):
         try:
             self.user_info = self.read_json_file(self.activation_flag)
 
-            logging.warning("{} already exists. Re-activation is not possible!".format(self.activation_flag))
-            logging.info("NuvlaBox credential: {}".format(self.user_info["api-key"]))
+            self.activate_logger.warning("{} already exists. Re-activation is not possible!".format(self.activation_flag))
+            self.activate_logger.info("NuvlaBox credential: {}".format(self.user_info["api-key"]))
             return False, self.user_info
         except FileNotFoundError:
             # file doesn't exist yet,
             # But maybe the API was provided via env?
             api_key, api_secret = self.get_api_keys()
             if api_key and api_secret:
-                logging.info(f'Found API key set in environment, with key value {api_key}')
+                self.activate_logger.info(f'Found API key set in environment, with key value {api_key}')
                 self.user_info = {
                     "api-key": api_key,
                     "secret-key": api_secret
@@ -66,7 +67,7 @@ class Activate(NuvlaBoxCommon.NuvlaBoxCommon):
     def activate(self):
         """ Makes the anonymous call to activate the NuvlaBox """
 
-        logging.info('Activating "{}"'.format(self.nuvlabox_id))
+        self.activate_logger.info('Activating "{}"'.format(self.nuvlabox_id))
 
         try:
             self.user_info = self.api()._cimi_post('{}/activate'.format(self.nuvlabox_id))
@@ -74,7 +75,8 @@ class Activate(NuvlaBoxCommon.NuvlaBoxCommon):
             self.shell_execute(["timeout", "3s", "/lib/systemd/systemd-timesyncd"])
             self.user_info = self.api()._cimi_post('{}/activate'.format(self.nuvlabox_id))
         except requests.exceptions.ConnectionError as conn_err:
-            logging.error("Can not reach out to Nuvla at {}. Error: {}".format(self.nuvla_endpoint, conn_err))
+            self.activate_logger.error("Can not reach out to Nuvla at {}. Error: {}"
+                                       .format(self.nuvla_endpoint, conn_err))
             raise
 
         # Flags that the activation has been done
@@ -90,12 +92,12 @@ class Activate(NuvlaBoxCommon.NuvlaBoxCommon):
         """
         context_file = "{}/{}".format(self.data_volume, self.context)
 
-        logging.info('Managing NB context file {}'.format(context_file))
+        self.activate_logger.info('Managing NB context file {}'.format(context_file))
 
         try:
             current_context = self.read_json_file(context_file)
         except (ValueError, FileNotFoundError):
-            logging.warning("Writing {} for the first time".format(context_file))
+            self.activate_logger.warning("Writing {} for the first time".format(context_file))
             current_context = {}
 
         self.write_json_to_file(context_file, nuvlabox_resource)
@@ -104,14 +106,17 @@ class Activate(NuvlaBoxCommon.NuvlaBoxCommon):
 
     def vpn_commission_if_needed(self, current_nb_resource: dict, old_nb_resource: dict):
         """
-        Checks if the VPN server ID has changed in the NB resource, and if so, asks for VPN commissioning
+        Checks if the VPN server ID has changed in the NB resource, and if so, asks for
+        VPN commissioning
 
         :param current_nb_resource: current NuvlaBox resource, from Nuvla
         :param old_nb_resource: old content of the NuvlaBox resource
         :return:
         """
-        if current_nb_resource.get("vpn-server-id") != old_nb_resource.get("vpn-server-id"):
-            logging.info(f'VPN Server ID has been added/changed in Nuvla: {current_nb_resource.get("vpn-server-id")}')
+        if current_nb_resource.get("vpn-server-id") != \
+                old_nb_resource.get("vpn-server-id"):
+            self.activate_logger.info(f'VPN Server ID has been added/changed in Nuvla: '
+                                      f'{current_nb_resource.get("vpn-server-id")}')
 
             self.commission_vpn()
